types: rename InversifyContainers to containers.✏️ types: rename InversifySymbols to symbols.✏️
