+++ src/main.ts
@@ -1,6 +1,6 @@
 import { DiscordBot } from './interfaces/DiscordBot';
-import { SYMBOLS } from './inversify/InversifySymbols';
-import { container } from './inversify/InversifyContainers';
+import { SYMBOLS } from './di/symbols';
+import { container } from './di/containers';
 
 const bot = container.get<DiscordBot>(SYMBOLS.DiscordBot);
 bot.connectBot();
